datascience/pyproject.toml
[project]
name = "datascience"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "adjusttext>=1.3.0",
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.7",
    "numpy>=2.3.5",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "pyyaml>=6.0.3",
    "ruff>=0.14.5",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.3",
    "seaborn>=0.13.2",
    "shapely>=2.1.2",
    "statsmodels>=0.14.5",
    "xgboost>=3.1.1",
]

[tool.setuptools.packages.find]
where = ["src"]      

# Required for pyTest to find the modules
[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

